refactor(creation): Extract SkillsCard sub-components (#301)
Break up SkillsCard.tsx (1,443 → 536 lines) by extracting:
- useSkillsCardHandlers: handler callbacks, data accessors, constants
- SkillsListSection: skill groups and individual skills list rendering
- SkillsModalsSection: all 8 modal dialog instantiations
- SkillGroupCard: group display component

Update test mocks to cover new direct-import module paths.
Update architecture docs with new component hierarchy.
